separated credits content into separate file
