+++ app/frontend/src/pages/CreditsPage/_CreditsPageCreditsCards.scss
