+++ app/frontend/src/pages/CreditsPage/_index.scss
@@ -1,0 +1,1 @@
+@forward "CreditsPageCreditsCards";
